[MIG] connector_ecommerce: Migration to 18.0
