+++ connector_ecommerce/models/__init__.py
@@ -1,6 +1,9 @@
-from . import account
-from . import invoice
-from . import account_payment_mode
-from . import product
-from . import sale
-from . import stock
+from . import account_tax
+from . import account_move
+from . import account_payment_method
+from . import account_payment_method_line
+from . import product_template
+from . import product_product
+from . import sale_order
+from . import stock_picking
+from . import stock_move
