+++ connector_ecommerce/models/stock_picking.py
@@ -36,26 +36,3 @@ class StockPicking(models.Model):
                 self._event("on_picking_dropship_done").notify(picking, method)
 
         return result
-
-
-class StockMove(models.Model):
-    _inherit = "stock.move"
-
-    def _action_done(self, cancel_backorder=False):
-        fire_event = not self.env.context.get("__no_on_event_out_done")
-        if fire_event:
-            pickings = self.mapped("picking_id")
-            states = {p.id: p.state for p in pickings}
-
-        result = super()._action_done(cancel_backorder=cancel_backorder)
-
-        if fire_event:
-            for picking in pickings:
-                if states[picking.id] != "done" and picking.state == "done":
-                    if picking.picking_type_id.code != "outgoing":
-                        continue
-                    # partial pickings are handled in
-                    # StockPicking.do_transfer()
-                    picking._event("on_picking_out_done").notify(picking, "complete")
-
-        return result
