--- connector_ecommerce/readme/CREDITS.md
@@ -1,2 +1,0 @@
-The migration of this module from 13.0 to 14.0 was financially supported
-by Camptocamp.
