+++ connector_ecommerce/static/description/index.html
@@ -409,8 +409,7 @@ Add structures shared for e-commerce connectors</blockquote>
 <li><a class="reference internal" href="#credits" id="toc-entry-3">Credits</a><ul>
 <li><a class="reference internal" href="#authors" id="toc-entry-4">Authors</a></li>
 <li><a class="reference internal" href="#contributors" id="toc-entry-5">Contributors</a></li>
-<li><a class="reference internal" href="#other-credits" id="toc-entry-6">Other credits</a></li>
-<li><a class="reference internal" href="#maintainers" id="toc-entry-7">Maintainers</a></li>
+<li><a class="reference internal" href="#maintainers" id="toc-entry-6">Maintainers</a></li>
 </ul>
 </li>
 </ul>
@@ -442,13 +441,8 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <p>See
 <a class="reference external" href="https://github.com/OCA/connector-ecommerce/graphs/contributors">contributors</a>.</p>
 </div>
-<div class="section" id="other-credits">
-<h2><a class="toc-backref" href="#toc-entry-6">Other credits</a></h2>
-<p>The migration of this module from 13.0 to 14.0 was financially supported
-by Camptocamp.</p>
-</div>
 <div class="section" id="maintainers">
-<h2><a class="toc-backref" href="#toc-entry-7">Maintainers</a></h2>
+<h2><a class="toc-backref" href="#toc-entry-6">Maintainers</a></h2>
 <p>This module is maintained by the OCA.</p>
 <a class="reference external image-reference" href="https://odoo-community.org">
 <img alt="Odoo Community Association" src="https://odoo-community.org/logo.png" />
