+++ connector_ecommerce/views/account_journal.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="view_account_journal_form" model="ir.ui.view">
+        <field name="model">account.journal</field>
+        <field name="inherit_id" ref="account.view_account_journal_form" />
+        <field name="arch" type="xml">
+            <xpath
+                expr="//field[@name='inbound_payment_method_line_ids']//field[@name='payment_account_id']"
+                position="after"
+            >
+                <field name="import_rule" />
+                <field name="days_before_cancel" />
+            </xpath>
+        </field>
+    </record>
+</odoo>
