--- connector_ecommerce/views/payment_mode_view.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="account_payment_mode_form_inherit" model="ir.ui.view">
-        <field name="name">account.payment.mode.connector_ecommerce.form</field>
-        <field name="model">account.payment.mode</field>
-        <field name="inherit_id" ref="account_payment_mode.account_payment_mode_form" />
-        <field name="arch" type="xml">
-            <group name="main" position="after">
-                <group name="import_rule" string="Import Rules">
-                    <group>
-                        <field name="import_rule" />
-                        <field name="days_before_cancel" />
-                    </group>
-                    <group name="import_rule_help" class="oe_grey" col="1">
-                        <p invisible="import_rule != 'never'">
-                            The sales orders using the payment method will never be imported.
-                        </p>
-                        <p invisible="import_rule != 'always'">
-                            The sales orders using the payment method will always be imported.
-                        </p>
-                        <p invisible="import_rule != 'paid'">
-                            The sales orders using the payment
-                            method will be imported
-                            only when they receive a payment on the E-Commerce backend.
-                            <br />
-                        </p>
-                    </group>
-                </group>
-            </group>
-        </field>
-    </record>
-    <record id="account_payment_mode_tree_inherit" model="ir.ui.view">
-        <field name="name">account.payment.mode.connector_ecommerce.tree</field>
-        <field name="model">account.payment.mode</field>
-        <field name="inherit_id" ref="account_payment_mode.account_payment_mode_tree" />
-        <field name="arch" type="xml">
-            <field name="company_id" position="before">
-                <field name="import_rule" />
-            </field>
-        </field>
-    </record>
-</odoo>
