+++ connector_ecommerce/views/sale_order.xml
@@ -1,31 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <record id="view_order_connector_form" model="ir.ui.view">
-        <field name="name">sale.order.connector.form</field>
+    <record id="sale_order_view_form" model="ir.ui.view">
         <field name="model">sale.order</field>
         <field name="inherit_id" ref="sale.view_order_form" />
         <field name="arch" type="xml">
-            <xpath expr="//notebook" position="inside">
-                <!-- change the invisible attribute to 0 when used in
-                    submodules -->
-                <page
-                    name="connector"
-                    string="Connectors"
-                    invisible="1"
-                    groups="connector.group_connector_manager"
-                />
-            </xpath>
-        </field>
-    </record>
-    <record id="sale_order_view_form" model="ir.ui.view">
-        <field name="name">sale_order.view_form</field>
-        <field name="model">sale.order</field>
-        <field name="inherit_id" ref="sale_automatic_workflow.sale_order_view_form" />
-        <field name="arch" type="xml">
-            <field name="workflow_process_id" position="after">
-                <field name="canceled_in_backend" />
-                <field name="cancellation_resolved" invisible="1" />
-            </field>
             <xpath expr="//header" position="inside">
                 <field name="parent_id" invisible="1" />
                 <button
@@ -36,8 +14,6 @@
                     invisible="not parent_id"
                     groups="base.group_user"
                 />
-            </xpath>
-            <xpath expr="//header" position="inside">
                 <button
                     name="%(action_sale_ignore_cancel)d"
                     type="action"
@@ -46,10 +22,25 @@
                     groups="sales_team.group_sale_manager"
                 />
             </xpath>
+            <xpath expr="//notebook" position="inside">
+                <!-- change the invisible attribute to 0 when used in
+                    submodules -->
+                <page
+                    name="connector"
+                    string="Connectors"
+                    invisible="1"
+                    groups="connector.group_connector_manager"
+                >
+                    <group>
+                        <field name="canceled_in_backend" />
+                        <field name="cancellation_resolved" invisible="1" />
+                    </group>
+                </page>
+            </xpath>
         </field>
     </record>
+
     <record id="view_sales_order_filter" model="ir.ui.view">
-        <field name="name">sale.order.search</field>
         <field name="model">sale.order</field>
         <field name="inherit_id" ref="sale.view_sales_order_filter" />
         <field name="arch" type="xml">
