+++ connector_ecommerce/wizard/sale_ignore_cancel_view.xml
@@ -6,28 +6,33 @@
         <field name="arch" type="xml">
             <form string="Ignore the cancellation on the Backend">
                 <p class="oe_grey">
-                        This sales order has been canceled from the backend.
-                        The usual action would be to cancel it in Odoo along
-                        all the documents generated (delivery orders, invoices, ...).
-                    </p>
+                    This sales order has been canceled from the backend.
+                    The usual action would be to cancel it in Odoo along
+                    all the documents generated (delivery orders, invoices, ...).
+                </p>
                 <p class="oe_grey">
-                        However, if for any reason you need to keep it open in Odoo,
-                        write the reason here and it will stay open.
-                    </p>
+                    However, if for any reason you need to keep it open in Odoo,
+                    write the reason here and it will stay open.
+                </p>
                 <group>
                     <label for="reason" colspan="2" />
                     <field name="reason" nolabel="1" colspan="2" />
                 </group>
                 <footer>
-                        <button
-                    name="confirm_ignore_cancel"
-                    string="Confirm"
-                    type="object"
-                    class="oe_highlight"
-                />
-                        or
-                        <button string="Cancel" class="oe_link" special="cancel" />
-                    </footer>
+                    <button
+                        name="confirm_ignore_cancel"
+                        string="Confirm"
+                        type="object"
+                        class="oe_highlight"
+                        data-hotkey="q"
+                    />
+                    <button
+                        string="Cancel"
+                        class="btn-secondary"
+                        special="cancel"
+                        data-hotkey="x"
+                    />
+                </footer>
             </form>
         </field>
     </record>
